requirements.txt
openai>=1.0.0
requests>=2.28.0
python-dotenv>=1.0.0
pandas>=1.5.0
matplotlib>=3.6.0
seaborn>=0.12.0
numpy>=1.23.0
plotly>=5.10.0
beautifulsoup4>=4.12.0
mkdocs>=1.4.0
mkdocs-material>=9.0.0
